Update WhatsNextService to throw exceptions if user is not authenticated
- we have to partially rely on the letterboxd service here because
although we can tell when the user is not signed in in some cases (e.g.
when we don't have a token at all), we don't know if they've changed
their credentials on the web, which _ought_ to invalidate our refresh
token, hence the runtime error.
- at the moment, we need the user to be signed in when they click to see
film details, but this is only because we get film relationship status
for the user. In reality, we could separate so we allow a non-signed in
experience and only ask them to sign in when they try to interact with a
film, but this is easier for now
- the observers for screens that require login should now check for
authentication errors.
- a simple way would be to create an observer that handles this error,
opening the sign-in screen, then forcing the user to retry their operation
by returning them to where they were
- if `signInActivity` was started with a startActivityForResult, then we
could instead retry loading the screen onActivityResult